+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.1] - 2021-12-23
+
+### Added
+- Add empty people indicator to events with no attendees
+
+### Fixed
+- Fix unpaid events with no attendees being categorized as paid
+- Fix invalid event progress warning indicator with no paid attendees
+
 ## [2.0.0] - 2021-12-22
 
 ### Added